Refactor tests a bit by adding InsertOneTableIntegrationTest, splitting from FindOne - test
