+++ src/test/java/io/stargate/sgv2/jsonapi/api/v1/tables/FindOneTableIntegrationTest.java
@@ -3,7 +3,6 @@ package io.stargate.sgv2.jsonapi.api.v1.tables;
 import io.quarkus.test.common.WithTestResource;
 import io.quarkus.test.junit.QuarkusIntegrationTest;
 import io.stargate.sgv2.jsonapi.api.v1.util.DataApiCommandSenders;
-import io.stargate.sgv2.jsonapi.exception.DocumentException;
 import io.stargate.sgv2.jsonapi.exception.ErrorCodeV1;
 import io.stargate.sgv2.jsonapi.testresource.DseTestResource;
 import java.util.Map;
@@ -19,8 +18,6 @@ import org.junit.jupiter.api.TestClassOrder;
 @TestClassOrder(ClassOrderer.OrderAnnotation.class)
 public class FindOneTableIntegrationTest extends AbstractTableIntegrationTestBase {
   static final String TABLE_WITH_STRING_ID_AGE_NAME = "findOneSingleStringKeyTable";
-  static final String TABLE_WITH_TEXT_COLUMNS = "findOneTextColumnsTable";
-  static final String TABLE_WITH_INT_COLUMNS = "findOneIntColumnsTable";
 
   @BeforeAll
   public final void createDefaultTables() {
@@ -33,32 +30,6 @@ public class FindOneTableIntegrationTest extends AbstractTableIntegrationTestBas
             Map.of("type", "int"),
             "name",
             Map.of("type", "text")),
-        "id");
-    createTableWithColumns(
-        TABLE_WITH_TEXT_COLUMNS,
-        Map.of(
-            "idText",
-            Map.of("type", "text"),
-            "asciiText",
-            Map.of("type", "ascii"),
-            "varcharText",
-            Map.of("type", "text")),
-        "idText");
-    createTableWithColumns(
-        TABLE_WITH_INT_COLUMNS,
-        Map.of(
-            "id",
-            Map.of("type", "text"),
-            "intValue",
-            Map.of("type", "int"),
-            "longValue",
-            Map.of("type", "bigint"),
-            "shortValue",
-            Map.of("type", "smallint"),
-            "byteValue",
-            Map.of("type", "tinyint"),
-            "bigIntegerValue",
-            Map.of("type", "varint")),
         "id");
   }
 
@@ -222,105 +193,6 @@ public class FindOneTableIntegrationTest extends AbstractTableIntegrationTestBas
 
   @Nested
   @Order(3)
-  class FindOneTextColumns {
-    public final String STRING_UTF8_WITH_2BYTE_CHAR = "utf8-2-byte-\u00a2"; // cent symbol
-    public final String STRING_UTF8_WITH_3BYTE_CHAR = "utf8-3-byte-\u20ac"; // euro symbol
-
-    @Test
-    void insertWithTextColumnsAndFind() {
-      final String DOC_JSON =
-              """
-                                  {
-                                      "idText": "abc",
-                                      "asciiText": "safe value",
-                                      "varcharText": "%s/%s"
-                                  }
-                                  """
-              .formatted(STRING_UTF8_WITH_2BYTE_CHAR, STRING_UTF8_WITH_3BYTE_CHAR);
-      insertOneInTable(TABLE_WITH_TEXT_COLUMNS, DOC_JSON);
-
-      DataApiCommandSenders.assertTableCommand(keyspaceName, TABLE_WITH_TEXT_COLUMNS)
-          .postFindOne("{ \"filter\": { \"idText\": \"abc\" } }")
-          .hasNoErrors()
-          .hasJSONField("data.document", DOC_JSON);
-    }
-
-    @Test
-    void failTryingToInsertNonAscii() {
-      final String DOC_JSON =
-              """
-                {
-                    "idText": "def",
-                    "asciiText": "%s",
-                    "varcharText": "safe value"
-                }
-                """
-              .formatted(STRING_UTF8_WITH_2BYTE_CHAR);
-      DataApiCommandSenders.assertTableCommand(keyspaceName, TABLE_WITH_TEXT_COLUMNS)
-          .postInsertOne(DOC_JSON)
-          .hasSingleApiError(
-              DocumentException.Code.INVALID_COLUMN_VALUES,
-              DocumentException.class,
-              "String contains non-ASCII character");
-    }
-  }
-
-  @Nested
-  @Order(4)
-  class FindOneIntColumns {
-    // [data-api#1429]: Test to verify that all-zero fractional parts are ok for int types
-    @Test
-    void insertWithIntColumnsZeroFractional() {
-      // In goes 5.00, out comes 5:
-      insertOneInTable(TABLE_WITH_INT_COLUMNS, numDoc("zero-fraction", "5.00"));
-      DataApiCommandSenders.assertTableCommand(keyspaceName, TABLE_WITH_INT_COLUMNS)
-          .postFindOne("{ \"filter\": { \"id\": \"zero-fraction\" } }")
-          .hasNoErrors()
-          .hasJSONField("data.document", numDoc("zero-fraction", "5"));
-    }
-
-    // [data-api#1429]: Test to verify that scientific is allowed for int types if  (and only if)
-    // the fractional part is zero
-    @Test
-    void insertWithIntColumnsScientificNotation() {
-      // In goes 1.23E+02, out comes 123:
-      insertOneInTable(TABLE_WITH_INT_COLUMNS, numDoc("scientific-but-int", "1.23E+02"));
-      DataApiCommandSenders.assertTableCommand(keyspaceName, TABLE_WITH_INT_COLUMNS)
-          .postFindOne("{ \"filter\": { \"id\": \"scientific-but-int\" } }")
-          .hasNoErrors()
-          .hasJSONField("data.document", numDoc("scientific-but-int", "123"));
-    }
-
-    // [data-api#1429]: Test to verify that should there be real fraction, insert fails
-    @Test
-    void failWithNonZeroFractionPlain() {
-      // Try with 12.5, should fail
-      DataApiCommandSenders.assertTableCommand(keyspaceName, TABLE_WITH_INT_COLUMNS)
-          .postInsertOne(numDoc("non-zero-fraction", "12.5"))
-          .hasSingleApiError(
-              DocumentException.Code.INVALID_COLUMN_VALUES,
-              DocumentException.class,
-              "Root cause: Rounding necessary");
-    }
-
-    private String numDoc(String id, String num) {
-      return
-          """
-                                  {
-                                      "id": "%s",
-                                      "byteValue": %s,
-                                      "shortValue": %s,
-                                      "intValue": %s,
-                                      "longValue": %s,
-                                      "bigIntegerValue": %s
-                                  }
-                                  """
-          .formatted(id, num, num, num, num, num);
-    }
-  }
-
-  @Nested
-  @Order(5)
   class FindOneFail {
     @Test
     @Order(1)
